refactor(healthz): seal Probe interface + sink ctxSafeProbe to kernel (#1187 review F6)
- Add unexported isHealthzProbe() marker to Probe interface — type-system Hard
  upstream gate; only funcProbe (kernel/healthz) and ctxSafeProbe (sunk from
  runtime/observability/healthz) can implement the interface
- Sink ctxSafeProbe / WrapCtxSafe / probeOutcome / watchLateOutcome / probeLogKey
  to new kernel/healthz/ctxsafe.go; runtime aggregator calls healthz.WrapCtxSafe
- Add kernel/healthz/probetest.go: NewTestOnlyProbeWithAnyName for conformance
  harness to exercise Aggregator empty-name validation without bypassing the seal
- Update healthztest conformance.go: replace emptyNameProbe{} (now uncompilable
  outside kernel/healthz) with NewTestOnlyProbeWithAnyName("", ...)
- Closes round-3 review F6 finding: Probe was an open interface allowing external
  types to bypass NewProbe funnel via Name() ProbeName + Check impl

Refs PR #1187 round-3 review
